fix: green CI — workspace fmt, clippy -D warnings clean, pnpm frontend, prebuilt sqlx-cli
- cargo fmt at workspace root (geo-engine was unformatted)
- clippy -D warnings clean: elided lifetimes, is_empty/div_ceil/contains
  idioms, underscore-prefixed validated-but-unemitted fields, allow(dead_code)
  on API contract structs, allow(too_many_arguments) on 9-arg instruction,
  UncheckedAccount for unchecked PDAs, crate-level allow(unexpected_cfgs)
  for Anchor macro expansion
- ci: prebuilt sqlx-cli 0.8.6 via install-action (latest 0.9.0 source build
  fails on tinyvec with current stable rustc)
- ci: frontend uses pnpm (pnpm-lock.yaml) on Node 22

+++ terra-core/api/src/routes/identities.rs
@@ -356,7 +356,7 @@ pub async fn request_succession(
         ));
     }
     let vc = req.validators.len() as u8;
-    if (required as u8) > vc {
+    if required > vc {
         return Err(AppError::bad_request(
             "required_validations cannot exceed the number of declared validators",
         ));
